chg: :truck: move dag inside a submodule
